+++ src/strings/show_typst.jl
@@ -49,22 +49,13 @@ function show_typst(io::IO, tc::TypstContext, x::Complex{<:Union{
 }})
     _mode = mode(tc)
     math_mode(io, tc, x) do io, tc, x
-        _real, _imag = real(x), imag(x)
-        _parenthesize = parenthesize(tc) && !(iszero(_real) || iszero(_imag))
         enclose(io, x, (_mode == math && _parenthesize ? ("(", ")") : ("", ""))...) do io, x
-            if iszero(_real) && !iszero(_imag)
-                signbit(_imag) && print(io, '-')
-                isone(abs(_imag)) || show_typst(io, abs(_imag))
-                print(io, 'i')
-            else
-                show_typst(io, _real)
+            imaginary = imag(x)
 
-                if !iszero(_imag)
-                    print(io, ' ', signbit(_imag) ? '-' : '+', ' ')
-                    isone(abs(_imag)) || show_typst(io, abs(_imag))
-                    print(io, 'i')
-                end
-            end
+            show_typst(io, real(x))
+            print(io, ' ', signbit(imaginary) ? '-' : '+', ' ')
+            show_typst(io, imaginary)
+            print(io, 'i')
         end
     end
 end
@@ -78,26 +69,6 @@ end
 show_typst(io::IO, ::TypstContext, x::Complex{<:Union{Bool, Unsigned}}) = show_typst(
     io, Complex(signed(real(x)), signed(imag(x)))
 )
-function show_typst(io::IO, typst_context::TypstContext, x::Expr)
-    (; head, args) = x
-
-    if head == :(=)
-        math_mode(io, typst_context, args) do io, _, args
-            show_typst(io, args[1])
-            print(io, " = ")
-            show_typst(io, args[2])
-        end
-    elseif head == :call show_typst(io, TypstFunction(args[1], Tuple(@view args[2:end])))
-    elseif head == :hcat show_typst(io, permutedims(args))
-    elseif head == :vect show_typst(io, args)
-    elseif head == :vcat
-        if all(arg -> arg isa Expr && arg.head == :row, args)
-            show_typst(io, stack(Iterators.map(arg -> arg.args, args); dims = 1))
-        else show_typst(io, args)
-        end
-    else error("`show_typst` is not yet implemented for an `Expr` with the head `$head`")
-    end
-end
 show_typst(io::IO, tc::TypstContext, x::HTML) = show_raw(io, tc, MIME"text/html"(), :html, x)
 show_typst(io::IO, tc::TypstContext, x::Irrational{T}) where T = math_mode(io, tc, x) do io, _, x
     if T isa Symbol && length(string(T)) == 1 print(io, x)